Refactor raspberry pi specific functions in their own file
